+++ truffle/src/com.oracle.truffle.api/src/com/oracle/truffle/api/frame/FrameDescriptor.java
@@ -61,8 +61,7 @@ import org.graalvm.collections.EconomicMap;
 
 /**
  * Descriptor of the slots of frame objects. Multiple frame instances are associated with one such
- * descriptor. The FrameDescriptor is not thread-safe until it's given to a first RootNode's
- * constructor. After that it has thread-safe properties.
+ * descriptor. The FrameDescriptor is thread-safe.
  *
  * @since 0.8 or earlier
  */
@@ -111,8 +110,8 @@ public final class FrameDescriptor implements Cloneable {
         this.defaultValue = defaultValue;
         this.slots = new ArrayList<>();
         this.identifierToSlotMap = EconomicMap.create();
-        this.version = createVersion();
         this.lock = lock == null ? this : lock;
+        newVersion(this);
     }
 
     /**
@@ -336,9 +335,9 @@ public final class FrameDescriptor implements Cloneable {
                  * First, only invalidate before updating kind so it's impossible to read a new kind
                  * and old still valid assumption.
                  */
-                version.invalidate();
+                invalidateVersion(this);
                 frameSlot.kind = kind;
-                version = createVersion();
+                newVersion(this);
             }
         }
     }
@@ -355,9 +354,8 @@ public final class FrameDescriptor implements Cloneable {
     }
 
     /**
-     * Returns the size of an array which is needed for storing all the slots in it using their
-     * {@link FrameSlot#getIndex()} as a position in the array. (The number may be bigger than the
-     * number of slots, if some slots are removed.)
+     * Returns the size of an array which is needed for storing all the frame slots. (The number may
+     * be bigger than the number of slots, if some slots are removed.)
      *
      * @return the size of the frame
      * @since 0.8 or earlier
@@ -490,8 +488,16 @@ public final class FrameDescriptor implements Cloneable {
      * Invalidates the current, and create a new version assumption.
      */
     private void updateVersion() {
-        version.invalidate();
-        version = createVersion();
+        invalidateVersion(this);
+        newVersion(this);
+    }
+
+    private static void newVersion(FrameDescriptor descriptor) {
+        descriptor.version = Truffle.getRuntime().createAssumption("frame version");
+    }
+
+    private static void invalidateVersion(FrameDescriptor descriptor) {
+        descriptor.version.invalidate();
     }
 
     /**
@@ -504,10 +510,6 @@ public final class FrameDescriptor implements Cloneable {
      */
     public Assumption getVersion() {
         return version;
-    }
-
-    private static Assumption createVersion() {
-        return Truffle.getRuntime().createAssumption("frame version");
     }
 
     /**
@@ -575,7 +577,7 @@ public final class FrameDescriptor implements Cloneable {
                 } else {
                     comma = true;
                 }
-                sb.append(slot.getIndex()).append(":").append(slot.getIdentifier());
+                sb.append(slot.index).append(":").append(slot.getIdentifier());
             }
             sb.append("}");
             return sb.toString();